+++ packages/kitchensink/service-logger/service-logger.test.js
@@ -1,0 +1,7 @@
+describe('service-logger', () => {
+  it('should log stuff', async () => {
+    const res = await get('/');
+    console.log(res);
+    // expect(res.data.countries.length).toBeGreaterThan(1);
+  });
+});