+++ packages/service-logger/src/hooks.js
@@ -1,0 +1,7 @@
+const SERVICE_NAME = `logger`;
+const LOGGER_TRANSPORTS = `${SERVICE_NAME}/transports`;
+
+module.exports = {
+  SERVICE_NAME,
+  LOGGER_TRANSPORTS,
+};