+++ packages/service-logger/src/lib/hooks.js
@@ -1,3 +1,0 @@
-import { SERVICE } from '@forrestjs/hooks'
-export const SERVICE_NAME = `${SERVICE} logger`
-export const LOGGER_TRANSPORTS = `${SERVICE_NAME}/transports`