+++ packages/service-logger/CHANGELOG.md
@@ -1,2 +1,8 @@
 # Change Log
 
+## v4.0.0
+
+- Remove Babel
+- Remove custom methods
+- Rename context's instance api to `log`
+- Improve fastify integration